Using iframe for tutor

+++ app/controllers/tutor_controller.rb
@@ -9,16 +9,17 @@ class TutorController < ApplicationController
         submission = Submission.find(params[:submission_id])
         authorize submission
 
-        file = File.join(DATA_DIR, params[:submission_id] + '-' + params[:id])
-        f = File.new(file, "w")
-        f.write(submission.code)
-        f.write("\n")
-        f.write("echo('I am Dieter')")
-        f.close
-        @codefile = file
-        render
-
-
-        File.delete(file)
+        if params.key?("fullscreen")
+            @fullscreen = true
+            file = File.join(DATA_DIR, params[:submission_id] + '-' + params[:id])
+            f = File.new(file, "w")
+            f.write(submission.code)
+            f.write("\n")
+            f.write("echo('I am Dieter')")
+            f.close
+            @codefile = file
+            render layout: false
+            File.delete(file)
+        end
     end
 end
